+++ packages/sdk-sui/src/index.ts
@@ -3,6 +3,7 @@ export * from './utils/createSuiClient';
 export * from './web3Sdk/claimLBTC';
 export * from './web3Sdk/getBalance';
 export * from './web3Sdk/getBasculeDepositStatus';
+export * from './web3Sdk/isBasculeCheckEnabled';
 export * from './web3Sdk/signLbtcDestionationAddrSui';
 export * from './web3Sdk/signTermsOfService';
 export * from './web3Sdk/unstakeLBTC';